apply cn server - operator done

+++ lib/screens/operator/detail/operator_skills/widgets/operator_skill_info_widget.dart
@@ -129,7 +129,7 @@ class _OperatorSkillInfoWidgetState extends State<OperatorSkillInfoWidget>
                         ),
                         SkillSpTypeWidget(
                           isSkillType: true,
-                          type: skill.skillType!.toInt(),
+                          type: skill.skillType!,
                         ),
                         if (skill.duration != null && skill.duration! > 0)
                           Container(
